Doppelganger buster helpers for the last ones of the Scientology era
I will give these ones for you for free because Scientology is not doing anything and everyone is threatening people here and there with the expense of my laser passport. These are intellectual property of mine originally but not anymore.  These ones will start spreading anyway like hell fire. That however does not make it so legal to use these besides you.

The microbiology hazard is an dangerous bacterial, partly dormant related to hell fire  people. Past bodies of humans in volcano holes that are almost full black, some white bird poo texture and icy slimy, perhaps a bit cold surface. The whole body is the bacteria but that is probably just a sign, never know. My holiday I without a valid reason booked to Santa Cruz in Gran Canary some months ago was paid by a random and I decided to not go because of Scientology will not scam me.  The Xenu something else started some days during those, but I was like the fuck, I am not a superhero in